POSIX arch fixes,csp_system and csp_services fixes
* Adding tasklist support via optional libprocps linkage

* csp_system.h: Found it odd that CSP_RPS_MTU exists but wasn't used as
  an upper board implicitly. csp_sys_tasklist() now takes a size
  argument that is set to CSP_RPS_MTU in csp_service_handler(). This
  change effected freertos code as well - but not logically.

* csp_thread.c: Normalizes pthread stack usage against known tested
  values, which is minimum of PTHREAD_STACK_MIN * 2. If the stack_size
  is set to 0, then getrlimit() is used to find the currently set
  process thread stack size and uses that - that value is usually very
  large, but it's the system's/process's default so it's fine.

* csp_thread.c: Threads are named on linux using pthread_setname_np()

All of the csp_thread changes help the new csp_sys_tasklist()
implementation print useful things. During a simple test, I issued a
csp_ps(8, 1000) and saw:

PS node 8:
testapp    198366   0.00%
ROUTER     198367   0.00%
SRV        198368   0.00%

This is pretty close the FreeRTOS output, which is nice.

There was a hardcoded "95" for csp_ps()'s expected packet size, which
was wierd since CSP_RPS_MTU exists. Again, none of these values are
dependent on the csp config for buffers, so having a hardcoded value
(even CSP_RPS_MTU) is bad. So a new header was created, csp_services.h
to expose this MTU to both the client and server of RPS.